Moves TaskQueueService to sync package
