+++ main/src/main/AndroidManifest.xml
@@ -61,7 +61,7 @@
 
         <service android:name="com.google.android.apps.muzei.wearable.WearableSourceUpdateService" />
 
-        <service android:name="com.google.android.apps.muzei.TaskQueueService" />
+        <service android:name="com.google.android.apps.muzei.sync.TaskQueueService" />
         <service
             android:name="com.google.android.apps.muzei.sync.DownloadArtworkJobService"
             android:permission="android.permission.BIND_JOB_SERVICE"
